refactor kotlin
