+++ src/main/kotlin/me/ryguy/ctfbot/modules/ctf/CTFGameStatsCommand.kt
@@ -1,8 +1,7 @@
-package me.ryguy.ctfbot.cmds
+package me.ryguy.ctfbot.modules.ctf
 
 import discord4j.core.`object`.entity.Message
-import me.ryguy.ctfbot.replyWithFailure
-import me.ryguy.ctfbot.modules.ctf.CTFDiscordOnlyCommand
+import me.ryguy.ctfbot.util.replyWithFailure
 import me.ryguy.ctfbot.util.*
 import org.jsoup.nodes.Document
 import reactor.core.publisher.Mono
